Rename asyncio module to avoid Py3.5 language keyword.
